Extracted streamed bytes and friends, ensured that CurlManager tests execute only with pycurl available
